requirements.txt
#
##### requirements.txt #####
#
##### Requirements without Version Specifiers #####
pytest
numpy
pandas
PyYAML
#